# Sort extensions during JSON serialization