This PR proposes to make JSON serialization deterministic, i.e., always produce the same result for the same input. This matches Go's behavior for built-in types, which notably will explicitly [sort](https://github.com/golang/go/blob/40b3c0e58a0ae8dec4684a009bf3806769e0fc41/src/encoding/json/encode.go#L762) fields within a map when serializing.

v2/event/event_marshal.go's `WriteJson`'s only non-deterministic behavior is an iteration over a map of extensions, which is performed in random order. This PR takes a similar [approach](https://github.com/golang/go/blob/40b3c0e58a0ae8dec4684a009bf3806769e0fc41/src/encoding/json/encode.go#L762) as `encoding/json`, without their logic for "resolving" key names (since extension names are always strings), and without the use of `reflect.Value`.

The motivation behind this change is a subtle dependence that the [Restate](https://restate.dev/) framework has on JSON ordering being deterministic. This framework allows [for code blocks](https://docs.restate.dev/develop/go/journaling-results#journaled-actions) that are non-deterministic (e.g., creating a CloudEvent `Event` with the current time), but then expects when retrying that the journal is "consistent" on replays before any errors. In affect, this means that normal code should be deterministic. This breaks down when a CloudEvent is passed as input to a service and serialized as JSON by default: even when the `Event` hasn't changed, the serialization will produce different bytes, causing the framework to believe the code is non-deterministic when everything was the same until the final serialization of a CloudEvent.

For this framework, it is easy to work around - we need to manually convert to JSON and normalize it (e.g., by converting to a map and then back to JSON, to make use of golang's deterministic JSON serialization of maps). It is an unfortunate extra step caused by the CloudEvent SDK's use of a non-deterministic serialization and, at least in this use, unexpected difference given the behavior of the standard library.