@embano1 2025-01-19
@sargas I'm not an expert with Restate, but looking at the Restate SDK/APIs why could one not wrap the creation/serialization of a CloudEvent into a `Run()` call which will be journaled and replayed. Or is the issue actually during the journaling and how Restate (de)serializes JSON? Can you provide a code example so we can discuss? 

I'm a bit wary introducing changes to our core marshaling logic since I don't yet understand 1/ the impact it has to other users (perhaps relying on the current non-deterministic marshaling behavior) and 2/ the performance impact of the proposed code change.