@sargas 2025-01-21
@embano1 
>  I'm not an expert with Restate, but looking at the Restate SDK/APIs why could one not wrap the creation/serialization of a CloudEvent into a Run() call which will be journaled and replayed. Or is the issue actually during the journaling and how Restate (de)serializes JSON?

The last part is the correct - Restate has an implicit assumption that the serialization (using `json.Marshal` by default) doesn't change as long as the inputs are the same, which is true for all types in Go by default unless a custom `MarshalJSON` does otherwise.

> Can you provide a code example so we can discuss?
```go
import (
	"context"
	"fmt"
	cloudevents "github.com/cloudevents/sdk-go/v2"
	"github.com/google/uuid"
	"log"
	"math/rand/v2"
	"time"

	restate "github.com/restatedev/sdk-go"
	"github.com/restatedev/sdk-go/server"
)

func main() {
	if err := server.NewRestate().
		Bind(restate.Reflect(FirstService{})).
		Bind(restate.Reflect(SecondService{})).
		Start(context.Background(), ":9080"); err != nil {
		log.Fatal(err)
	}
}

type SecondService struct{}

func (SecondService) Process(_ restate.Context, _ cloudevents.Event) error {
	return nil
}

type FirstService struct{}

func (FirstService) Handle(ctx restate.Context) error {
	event, _ := restate.Run(ctx, func(ctx restate.RunContext) (cloudevents.Event, error) {
		event := cloudevents.NewEvent()
		event.SetID(uuid.NewString())
		event.SetSource("example/uri")
		event.SetType("example.type")
		event.SetTime(time.Now())
		event.SetExtension("a", "b")
		event.SetExtension("c", "d")
		event.SetExtension("e", "f")
		return event, nil
	})

	_, _ = restate.Service[restate.Void](ctx, "SecondService", "Process").Request(event)
	if rand.Float32() < 0.5 {
		return nil
	} else {
		return fmt.Errorf("failing with purposeful error")
	}
}
```
Requests to trigger FirstService/Handle will create the `cloudevents.Event` (inside `restate.Run`, since the current time is being used), pass it to the second service, and half the time go on happily.
Otherwise, if we randomly hit the error condition at the end, then the `FirstService.Handle` method should be called again - the `cloudevents.Event` from the first run will be returned from the `restate.Run` line from the journal, and the same (ignored) null response should be returned from the `restate.Service[..](..).Request` as that call is the next one in the journal. This may retry any number of times depending on the random number generator, but we shouldn't ever create more than one cloud event value or call the `SecondService` more than once.

The issue here is that since `event` serializes randomly when extensions are present, the journal entries created during retries can be inconsistent, leading to an error like:
```
2025/01/20 22:30:16 ERROR Journal mismatch: Replayed journal entries did not correspond to the user code. The user code has to be deterministic! method=FirstService/Handle invocationID=inv_1gdbLDeS59C52GA3G1nv6lPHPgBoN7LaHn expectedType=*wire.CallEntryMessage expectedMessage="{\"service_name\":\"SecondService\",\"handler_name\":\"Process\",\"parameter\":\"eyJzcGVjdmVyc2lvbiI6IjEuMCIsImlkIjoiMmMzNzJlY2YtNzY2Zi00MzcwLWI2YmMtZGIxNTE1M2M2OTlkIiwic291cmNlIjoiZXhhbXBsZS91cmkiLCJ0eXBlIjoiZXhhbXBsZS50eXBlIiwidGltZSI6IjIwMjUtMDEtMjFUMDQ6MzA6MTYuNTc0NTgxNzM3WiIsImUiOiJmIiwiYSI6ImIiLCJjIjoiZCJ9\",\"Result\":null}" actualType=*wire.CallEntryMessage actualMessage="{\"service_name\":\"SecondService\",\"handler_name\":\"Process\",\"parameter\":\"eyJzcGVjdmVyc2lvbiI6IjEuMCIsImlkIjoiMmMzNzJlY2YtNzY2Zi00MzcwLWI2YmMtZGIxNTE1M2M2OTlkIiwic291cmNlIjoiZXhhbXBsZS91cmkiLCJ0eXBlIjoiZXhhbXBsZS50eXBlIiwidGltZSI6IjIwMjUtMDEtMjFUMDQ6MzA6MTYuNTc0NTgxNzM3WiIsImEiOiJiIiwiYyI6ImQiLCJlIjoiZiJ9\",\"Result\":{\"Value\":\"\"}}"
```
Base64 decoding the "parameter" in the actual and expected journal messages shows they only different in the order of the extensions in the JSON-serialized form of the event.

> I'm a bit wary introducing changes to our core marshaling logic
I'm completely sympathetic to being cautious here - that's part of why I worded this PR as a proposal. I don't think there's anything wrong with the existing code (a Marshaler has to [create valid JSON](https://pkg.go.dev/encoding/json#Marshaler), but it isn't documented that it needs to be deterministic). My example above can have workarounds added, like dealing with strings or []byte's instead of `cloudevents.Event` or converting back and forth to a map to get a sort of normalized JSON.

At the same time, something is off about `cloudevents.Event` - it's rare to see a type that doesn't have deterministic marshaling. Restate is the first place I've seen that is affected by this, but given how widespread this convention is, it may not be the last piece of code with an implicit assumption that objects have a stable serialization.

> the impact it has to other users (perhaps relying on the current non-deterministic marshaling behavior)
I'm not sure anyone would _expect_ this, let alone rely on it. I'd expect the opposite is more common (people or software like Restate that assumes it would be deterministic).

> the performance impact of the proposed code change.
I don't have an answer to this, but I'll note that this PR still keeps the marshaling simpler then the code used in `encoding/json` for marshaling maps (since we have string field names). This approach of creating a slice for sorting the names is what the standard library uses anytime the JSON of a map is determined.