@duglin 2025-01-21
My perspective...
- if it's random today then code receiving CEs from our SDK it MUST be able to handle them being determinate going forward. I have no idea how it couldn't since it could (randomly) always end up being alphabetized
- performance: if someone has so many CE properties that this one "sort" impacts them I suspect they're sending too much data as CE properties and I this sort probably isn't their only performance concern. CE props are not meant to include all event data, just what's needed to help get it to it's final destination... where they crack open the body
- I'm anal enough that this one made me smile. I'm so bad that in my xRegistry impl I not only sort things I group the output so that system defined props come first (in order of importance to the user), then a blank line, then sorted extensions. My code doesn't depend on incoming stuff to be this way, but I wanted my output to be "human friendly" and deterministic for internal test validation purposes. If someone ever changes the code so that the output changes (even by one byte) I want that change to be scrutinized because (to me) UX is critical. So, I'm pretty biased here :-)  

Now... if someone were to suggest that we demand incoming CE's be sorted.... then that's a different story, obviously.