@duglin 2025-01-21
I will add though... I think Restate is wrong in its processing and in MOST other cases I would tell people to yell at them to fix their code because as @embano1 said... the JSON spec doesn't mandate that things be ordered. I'm just willing to let this ordering happen because I see a potential benefit and no harm to it.