@embano1 2025-01-21
> Like some kind of static global? That might mess up users' code and tests more. The Marshaler interface doesn't give any flexibility here for customizing the marshaling unfortunately

Before you wrote this, I also dug into the current implementation and I thought for extensions we'd use `encoding/json` directly instead of the `stream...` package. I better understand the issue now on your side.

I ran the existing benchmarks to get a better understanding on the impact:


```console
goos: darwin
goarch: arm64
pkg: github.com/cloudevents/sdk-go/v2/event
cpu: Apple M1 Pro
                                                                                   │ current.txt  │               new.txt               │
                                                                                   │    sec/op    │    sec/op     vs base               │
Marshal/struct_data_v0.3-10                                                          3.553µ ±  4%   3.611µ ±  7%       ~ (p=0.247 n=10)
Marshal/nil_data_v0.3-10                                                             3.308µ ±  7%   3.126µ ±  6%  -5.52% (p=0.003 n=10)
Marshal/string_data_v0.3-10                                                          3.544µ ±  4%   3.659µ ±  7%       ~ (p=0.210 n=10)
Marshal/struct_data_v1.0-10                                                          3.553µ ±  4%   3.557µ ±  4%       ~ (p=0.853 n=10)
Marshal/nil_data_v1.0-10                                                             3.345µ ± 15%   3.588µ ± 11%       ~ (p=0.143 n=10)
Marshal/string_data_v1.0-10                                                          3.406µ ±  2%   3.603µ ±  6%  +5.77% (p=0.019 n=10)
Marshal/base64_json_encoded_data_v1.0-10                                             1.895µ ±  3%   1.850µ ±  6%       ~ (p=0.119 n=10)
Marshal/number_data_v1.0-10                                                          1.960µ ±  8%   1.870µ ±  1%  -4.57% (p=0.000 n=10)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-10   4.359µ ±  5%   4.366µ ±  2%       ~ (p=0.837 n=10)
Unmarshal/string_data_v1.0-10                                                        4.202µ ±  1%   4.363µ ± 14%  +3.83% (p=0.000 n=10)
Unmarshal/nil_data_v1.0-10                                                           4.058µ ±  4%   4.067µ ± 10%       ~ (p=0.353 n=10)
Unmarshal/struct_data_v0.3-10                                                        4.343µ ± 15%   4.368µ ±  9%       ~ (p=0.631 n=10)
Unmarshal/nil_data_v0.3-10                                                           3.812µ ±  4%   3.884µ ±  2%       ~ (p=0.109 n=10)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-10   4.537µ ±  9%   4.609µ ±  5%       ~ (p=0.469 n=10)
Unmarshal/base64_json_encoded_data_v1.0-10                                           2.725µ ±  4%   2.695µ ±  2%       ~ (p=0.325 n=10)
Unmarshal/base64_xml_encoded_data_v1.0-10                                            2.837µ ±  4%   2.794µ ±  1%  -1.52% (p=0.003 n=10)
Unmarshal/string_data_v0.3-10                                                        4.203µ ±  5%   4.127µ ±  4%       ~ (p=0.143 n=10)
Unmarshal/struct_data_v1.0-10                                                        4.600µ ± 12%   4.338µ ±  3%  -5.70% (p=0.014 n=10)
geomean                                                                              3.466µ         3.471µ        +0.15%

                                                                                   │ current.txt  │                new.txt                 │
                                                                                   │     B/op     │     B/op      vs base                  │
Marshal/struct_data_v0.3-10                                                          2.228Ki ± 0%   2.465Ki ± 0%  +10.65% (p=0.000 n=10)
Marshal/nil_data_v0.3-10                                                             2.246Ki ± 0%   1.995Ki ± 0%  -11.15% (p=0.000 n=10)
Marshal/string_data_v0.3-10                                                          2.162Ki ± 0%   2.375Ki ± 0%   +9.85% (p=0.000 n=10)
Marshal/struct_data_v1.0-10                                                          2.255Ki ± 0%   2.465Ki ± 0%   +9.31% (p=0.000 n=10)
Marshal/nil_data_v1.0-10                                                             2.247Ki ± 0%   1.995Ki ± 0%  -11.21% (p=0.000 n=10)
Marshal/string_data_v1.0-10                                                          2.216Ki ± 0%   2.418Ki ± 0%   +9.12% (p=0.000 n=10)
Marshal/base64_json_encoded_data_v1.0-10                                               801.0 ± 0%     801.0 ± 0%        ~ (p=1.000 n=10) ¹
Marshal/number_data_v1.0-10                                                          1.563Ki ± 0%   1.563Ki ± 0%        ~ (p=1.000 n=10)
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-10   1.758Ki ± 0%   1.758Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/string_data_v1.0-10                                                        1.719Ki ± 0%   1.719Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/nil_data_v1.0-10                                                           1.680Ki ± 0%   1.680Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/struct_data_v0.3-10                                                        1.734Ki ± 0%   1.734Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/nil_data_v0.3-10                                                           1.680Ki ± 0%   1.680Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-10   1.758Ki ± 0%   1.758Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/base64_json_encoded_data_v1.0-10                                           1.203Ki ± 0%   1.203Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/base64_xml_encoded_data_v1.0-10                                            1.227Ki ± 0%   1.227Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/string_data_v0.3-10                                                        1.719Ki ± 0%   1.719Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/struct_data_v1.0-10                                                        1.734Ki ± 0%   1.734Ki ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                                              1.718Ki        1.731Ki        +0.75%
¹ all samples are equal

                                                                                   │ current.txt │               new.txt                │
                                                                                   │  allocs/op  │ allocs/op   vs base                  │
Marshal/struct_data_v0.3-10                                                           19.00 ± 0%   21.00 ± 0%  +10.53% (p=0.000 n=10)
Marshal/nil_data_v0.3-10                                                              17.00 ± 0%   18.00 ± 0%   +5.88% (p=0.000 n=10)
Marshal/string_data_v0.3-10                                                           20.00 ± 0%   23.00 ± 0%  +15.00% (p=0.000 n=10)
Marshal/struct_data_v1.0-10                                                           19.00 ± 0%   21.00 ± 0%  +10.53% (p=0.000 n=10)
Marshal/nil_data_v1.0-10                                                              17.00 ± 0%   18.00 ± 0%   +5.88% (p=0.000 n=10)
Marshal/string_data_v1.0-10                                                           19.00 ± 0%   21.00 ± 0%  +10.53% (p=0.000 n=10)
Marshal/base64_json_encoded_data_v1.0-10                                              9.000 ± 0%   9.000 ± 0%        ~ (p=1.000 n=10) ¹
Marshal/number_data_v1.0-10                                                           11.00 ± 0%   11.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v0.3-10    44.00 ± 0%   44.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/string_data_v1.0-10                                                         42.00 ± 0%   42.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/nil_data_v1.0-10                                                            40.00 ± 0%   40.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/struct_data_v0.3-10                                                         43.00 ± 0%   43.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/nil_data_v0.3-10                                                            40.00 ± 0%   40.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/data,_attributes_and_extensions_and_specversion_with_struct_data_v1.0-10    44.00 ± 0%   44.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/base64_json_encoded_data_v1.0-10                                            25.00 ± 0%   25.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/base64_xml_encoded_data_v1.0-10                                             25.00 ± 0%   25.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/string_data_v0.3-10                                                         42.00 ± 0%   42.00 ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/struct_data_v1.0-10                                                         43.00 ± 0%   43.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                                               25.75        26.56        +3.13%
¹ all samples are equal

```